move function to producer module
